refactor: move script to esm
